# wfzo-api-main/wfzo-api-main/api-spec.yml
openapi: 3.0.4
info:
  title: WFZO Website Management Platform API
  description: >-
    REST APIs for the World Free Zones Organization (WFZO) web platform.
    Scopes
     - Auth: Microsoft Entra (External ID) Bearer JWT
     - Content: read-through to Strapi CMS
     - Documents: secure PDF delivery via short-lived Azure Blob SAS
     - Events: Zoom registration with per-membership seat quotas
     - Learning: LMS integration with Moodle
     - Community: Discourse integrations (DMs, topics, embeds)
     - CMS APIs
  version: 1.0.0
servers:
  - url: /wfzo/api/v1
...
paths:
...
  /member:
    post:
      tags:
        - Member
      summary: Create a new member
      description: Create a new member record; supports basic registration and full onboarding.
      operationId: registration
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInfo"
      responses:
        "200":
          description: successful operation need to respond the memberId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
  /member/me:
    post:
      tags:
        - Member
      summary: Get member details by email
      description: Retrieve member information using the user's email address.
      operationId: getMemberByEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  description: User email address to find the member
                  example: "john.doe@example.com"
      responses:
        "200":
          description: Member details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberInfo"
        "400":
          description: Invalid input
        "404":
          description: Member not found
        default:
          description: Unexpected error
    get:
      tags:
        - Member
      summary: Search members
      description: Search members by various filters (e.g., featured, country, industry, starting letter, free text). Supports pagination.
      operationId: getMembers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberSearchResponse"
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        default:
          description: Unexpected error
  /member/activity:
    patch:
      tags:
        - Member
      summary: Log Member activity
      description: Record a member's feature usage (e.g., a click). The backend increments usage and responds with the latest count for that feature.
      operationId: logMemberActivity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberActivity"
            examples:
              Library Download:
                summary: Member downloaded a library document; server will increment the count
                value:
                  memberId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                  featureKey: "members.libraryDownloads"
                  timestamp: "2025-10-10T10:15:30Z"
              Member News Access:
                summary: Member accessed a Member News; server will increment the count
                value:
                  memberId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                  featureKey: "members.newsAccess"
                  timestamp: "2025-10-10T10:15:30Z"
      responses:
        "200":
          description: Activity logged successfully; returns the latest usage count after update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberActivity"
              examples:
                Library Download increment:
                  summary: Latest count after increment
                  value:
                    memberId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                    featureKey: "members.libraryDownloads"
                    usageCount: 11
                    timestamp: "2025-10-10T10:15:30Z"
                Member News Access increment:
                  summary: Latest count after increment
                  value:
                    memberId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                    featureKey: "members.newsAccess"
                    usageCount: 11
                    timestamp: "2025-10-10T10:15:30Z"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              examples:
                ERR005:
                  value:
                    code: "ERR005"
                    type: "FAILURE"
                    message: "Invalid activity data"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: "GEN001"
                type: "FAILURE"
                message: "Internal Server Error"
  /member/{memberId}:
    get:
      tags:
        - Member
      summary: Get member by ID
      description: Return the member profile and related membership information for the given member ID.
      operationId: getMemberById
      parameters:
        - name: memberId
          in: path
          description: Unique identifier of the member
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        default:
          description: Unexpected error
    put:
      tags:
        - Member
      summary: Update member details
      description: Update profile or membership information for the specified member.
      operationId: memberUpdate
      parameters:
        - name: memberId
          in: path
          description: Unique identifier of the member
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          application/xml:
            schema:
              $ref: "#/components/schemas/User"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
  /member/industries:
    get:
      tags:
        - Member
      summary: List industries
      description: Return the list of industries used for member classification and filtering.
      operationId: getIndustries
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                - "Aerospace"
                - "Agriculture"
                - "Automotive"
                - "Biotechnology"
                - "Chemicals"
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        default:
          description: Unexpected error
  /member/partnersandSponsors:
    get:
      tags:
        - Member
      summary: List partners and sponsors
      description: Return the current list of partnering organizations and sponsors with logos and links.
      operationId: getPartnersAndSponsors
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnersAndSponsors"
              example:
                partners:
                  - name: "Partner 1"
                    logoUrl: "https://example.com/logos/partner1.png"
                    websiteUrl: "https://partner1.com"
                  - name: "Partner 2"
                    logoUrl: "https://example.com/logos/partner2.png"
                    websiteUrl: "https://partner2.com"
                sponsors:
                  - name: "Sponsor 1"
                    logoUrl: "https://example.com/logos/sponsor1.png"
                    websiteUrl: "https://sponsor1.com"
                  - name: "Sponsor 2"
                    logoUrl: "https://example.com/logos/sponsor2.png"
                    websiteUrl: "https://sponsor2.com"
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        default:
          description: Unexpected error
  /member/featured:
    get:
      tags:
        - Member
      summary: List featured members
      description: Return a curated list of featured members to highlight on the site.
      operationId: getFeaturedMembers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeaturedMember"
              example:
                - memberId: "a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6"
                  name: "Member One"
                  logoUrl: "https://example.com/logos/member1.png"
                  websiteUrl: "https://memberone.com"
                - memberId: "p0o9i8u7-y6t5-r4e3-w2q1-z0x9c8v7b6n5"
                  name: "Member Two"
                  logoUrl: "https://example.com/logos/member2.png"
                  websiteUrl: "https://membertwo.com"
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        default:
          description: Unexpected error
...
  /member/status/{memberId}/{status}:
    put:
      tags:
        - Member
      summary: Update member status
      description: Update the membership status and related flags for the specified member ID.
      operationId: updateMemberStatus
      parameters:
        - name: memberId
          in: path
          description: Member identifier
          required: true
          schema:
            type: string
        - name: status
          in: path
          description: New membership status
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
...
components:
...
  schemas:
...
    Tier:
      type: array
      items:
        type: string
      enum: [basic, core, premium]
      description: Membership tier options (basic/core/premium)
    MembershipCategory:
      type: string
      description: Membership category options
      enum:
        [zoneMember, freeZoneAssociation, corporateMembers, professionalMembers, strategicMembers]
...
    MemberSearchResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          description: Wrapper including status and the member search payload
          properties:
            data:
              $ref: "#/components/schemas/MemberSearchData"
...
    MemberInfo:
      type: object
      description: Membership attributes and status for a member organization
      required: [category, status]
      properties:
        id:
          type: string
          format: uuid
        users:
          type: array
          items: { $ref: "#/components/schemas/User" }
        Address:
          $ref: "#/components/schemas/Address"
        category:
          type: string
          $ref: "#/components/schemas/MembershipCategory"
          description:
            Membership category. This is the basic Membership level. Current system supports 3
            categories votingMember, associateMember, partnerObserver with no additional tier. This is expected
            to be extended to 6 with multiple tier. We will also control the categories to be displayed or available
            in the website using a flag(active/inactive).
        tier:
          type: string
          $ref: "#/components/schemas/Tier"
          description:
            Membership tier within the category. This is expected to be extended to 3 tiers - basic, core, premium.
            We will also control the tiers to be displayed or available in the
        status:
          type: string
          enum: [active, expired, pendingApproval, pendingFormSubmission]
        validUntil:
          type: string
          format: date-time
          nullable: true
        OrganisationInfo:
          $ref: "#/components/schemas/OrganisationInfo"
        MemberConsent:
          $ref: "#/components/schemas/MemberConsent"
        featuredMember:
          type: boolean
          description: Whether to feature this member on the website
...
    ApiResponse:
      type: object
      description: Standard API response envelope for status reporting
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header